bind CCMenu, CCMenuItems

+++ tools/tolua++/oKeyFrame.h
@@ -5,7 +5,7 @@ CCActionInterval* oKeyScale::create @ oScale(float duration, float scaleX, float
 
 CCActionInterval* oKeyRotate::create @ oRotate(float duration, float rotate, unsigned char easeId = 0);
 
-CCActionInterval* oKeyOpacity::create @ oOpacity(float duration, float opacity, unsigned char easeId = 0);
+CCActionInterval* oKeyOpacity::create @ oOpacity(float duration, GLubyte opacity, unsigned char easeId = 0);
 
 CCActionInterval* oKeySkew::create @ oSkew(float duration, float skewX, float skewY, unsigned char easeId = 0);
